YT-19719: Fix test_attached_mode_op_aborted failures

+++ yt/python/yt/wrapper/default_config.py
@@ -53,7 +53,7 @@ def retry_backoff_config(**kwargs):
 
 def retries_config(**kwargs):
     if "total_timeout" in kwargs and isinstance(kwargs["total_timeout"], timedelta):
-        kwargs["total_timeout"] = int(kwargs["total_timeout"].total_seconds() * 1000)
+        kwargs["total_timeout"] = kwargs["total_timeout"].total_seconds() * 1000.0
 
     config_dict = {
         "count": None,
